requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
Django==3.2.10
django-cors-headers==4.4.0
djangorestframework==3.13.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
idna==3.7
jieba3k==0.35.1
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.3.0
lxml_html_clean==0.2.2
mysqlclient==2.2.4
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
packaging==24.0
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.31.0
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.11.0
urllib3==2.2.1
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
Django==3.2.10
django-cors-headers==4.4.0
djangorestframework==3.13.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
idna==3.7
jieba3k==0.35.1
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.3.0
lxml_html_clean==0.2.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
packaging==24.0
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.31.0
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.11.0
urllib3==2.2.1
